+++ lib/td_auth/ldap/ldap.ex
@@ -6,6 +6,8 @@ defmodule TdAuth.Ldap.Ldap do
 
   require Logger
 
+  @exldap_module Application.compile_env(:td_auth, :exldap_module)
+
   def authenticate(user_name, password) do
     case ldap_authenticate() do
       :ok -> create_profile(user_name, password)
@@ -29,15 +31,20 @@ defmodule TdAuth.Ldap.Ldap do
   end
 
   defp ldap_open do
-    Exldap.open(get_ldap_server(), get_ldap_port(), get_ldap_ssl(), get_ldap_connection_timeout())
+    @exldap_module.open(
+      get_ldap_server(),
+      get_ldap_port(),
+      get_ldap_ssl(),
+      get_ldap_connection_timeout()
+    )
   end
 
   defp ldap_close(conn) do
-    Exldap.close(conn)
+    @exldap_module.close(conn)
   end
 
   defp ldap_verify(conn) do
-    Exldap.verify_credentials(conn, get_ldap_user_dn(), get_ldap_password())
+    @exldap_module.verify_credentials(conn, get_ldap_user_dn(), get_ldap_password())
   end
 
   defp create_profile(user_name, password) do
@@ -75,11 +82,11 @@ defmodule TdAuth.Ldap.Ldap do
 
   defp verify_credentials(conn, user_name, password, nil) do
     bind = get_ldap_bind(user_name)
-    Exldap.verify_credentials(conn, bind, password)
+    @exldap_module.verify_credentials(conn, bind, password)
   end
 
   defp verify_credentials(conn, _user_name, password, object_name) do
-    Exldap.verify_credentials(conn, object_name, password)
+    @exldap_module.verify_credentials(conn, object_name, password)
   end
 
   defp get_ldap_bind(user_name) do
@@ -91,7 +98,7 @@ defmodule TdAuth.Ldap.Ldap do
   end
 
   defp ldap_connect do
-    Exldap.connect(
+    @exldap_module.connect(
       get_ldap_server(),
       get_ldap_port(),
       get_ldap_ssl(),
@@ -102,7 +109,7 @@ defmodule TdAuth.Ldap.Ldap do
   end
 
   defp ldap_search(conn, user_name) do
-    Exldap.search_field(conn, get_ldap_search_path(), get_ldap_search_field(), user_name)
+    @exldap_module.search_field(conn, get_ldap_search_path(), get_ldap_search_field(), user_name)
   end
 
   defp build_profile(entry) do
@@ -112,6 +119,7 @@ defmodule TdAuth.Ldap.Ldap do
       attr = Exldap.get_attribute!(entry, v)
       Map.put(acc, k, attr)
     end)
+    |> maybe_put_groups(entry, Application.get_env(:td_auth, :ldap)[:create_groups])
   end
 
   defp fetch_ldap_entry(search_results) do
@@ -159,4 +167,22 @@ defmodule TdAuth.Ldap.Ldap do
     timeout = Application.get_env(:td_auth, :ldap)[:connection_timeout]
     String.to_integer(timeout)
   end
+
+  defp maybe_put_groups(profile, %{attributes: entryAttributes}, true) do
+    fields = Application.get_env(:td_auth, :ldap)[:group_fields]
+
+    allowed_groups = Application.get_env(:td_auth, :ldap)[:allowed_groups]
+
+    profile_groups =
+      entryAttributes
+      |> Enum.filter(fn {key, _} -> Enum.member?(fields, to_string(key)) end)
+      |> Enum.flat_map(fn {_, value} -> value end)
+      |> Enum.map(&to_string/1)
+      |> Enum.uniq()
+      |> Enum.filter(&Enum.member?(allowed_groups, &1))
+
+    Map.put(profile, :groups, profile_groups)
+  end
+
+  defp maybe_put_groups(profile, _, _), do: profile
 end